+++ apps/manifest-demo/3008-webpack-host/remotes.d.ts
@@ -1,0 +1,6 @@
+declare module 'remote1/useCustomRemoteHook';
+declare module 'remote1/WebpackSvg';
+declare module 'remote1/WebpackPng';
+declare module 'manifest-provider/Component';
+declare module 'js-entry-provider/Component';
+declare module 'dynamic-remote/ButtonOldAnt';