+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @module-federation/runtime
 
+## 0.0.17
+
+### Patch Changes
+
+- ce0597e: feat: add registerRemotes api
+  - @module-federation/sdk@0.0.17
+
+## 0.0.16
+
+### Patch Changes
+
+- @module-federation/sdk@0.0.16
+
+## 0.0.15
+
+### Patch Changes
+
+- 6e9b6d5: fix(runtime): preserve error.stack instead of throwing new error
+- Updated dependencies [3a45d99]
+- Updated dependencies [ba5bedd]
+  - @module-federation/sdk@0.0.15
+
+## 0.0.14
+
+### Patch Changes
+
+- a050645: Expose node script loaders to bundler runtime. Replace require.loadScript from federation/node to use federation.runtime.loadScriptNode
+- 4fc20cc: adding reject to args in loadEsmEntry
+- Updated dependencies [cd8c7bf]
+- Updated dependencies [5576c6b]
+  - @module-federation/sdk@0.0.14
+
 ## 0.0.13
 
 ### Patch Changes